### 2019-07-25 #1594 @whedon
[PRE REVIEW]: ASCENDS: Advanced data SCiENce toolkit for Non-Data Scientists
**Submitting author:** @ornlpmcp (<a href="http://orcid.org/0000-0002-1317-5112">Sangkeun Lee</a>)
**Repository:** <a href="https://github.com/ornlpmcp/ASCENDS" target ="_blank">https://github.com/ornlpmcp/ASCENDS</a>
**Version:** v0.4.0
