### 2019-07-25 7b02835
Fixes ORCID in paper.md
