### 2019-07-25 #1594 @kyleniemeyer
Hi @ornlpmcp, thanks for your submission to JOSS. Please merge the PR I just submitted, which fixes an issue with an ORCID preventing compilation of your paper: https://github.com/ornlpmcp/ASCENDS/pull/1
