### 2019-07-30 b19c205
Update paper.md
