### 2019-07-30 274d6fe
Update paper.md
