### 2019-08-17 #1656 @whedon
[REVIEW]: ASCENDS: Advanced data SCiENce toolkit for Non-Data Scientists
**Submitting author:** @ornlpmcp (<a href="http://orcid.org/0000-0002-1317-5112">Sangkeun Lee</a>)
**Repository:** <a href="https://github.com/ornlpmcp/ASCENDS" target ="_blank">https://github.com/ornlpmcp/ASCENDS</a>
**Version:** 0.4.1
