### 2019-08-17 #1656 @terrytangyuan
@ornlpmcp Could you fix the compilation failure of the paper?
