### 2019-08-19 #1656 @zhampel
@ornlpmcp Here is my initial review. I will await your reply to my comments to continue my review.

**Paper**

	I think the paper needs further motivation to support your submission. Perhaps by describing why your package facilitates analysis either over existing packages or by highlighting its strengths. This would also help clarify the intended use case and audience. In addition to corrections to the text, I have made a few specific comments regarding this in the points below.

	- These success stories have been motivating scientists, who study physics, chemistry, materials, medicine and many more, —> These success stories have been motivating scientists who study physics, chemistry, materials, medicine and many other subjects,

	- Remove ‘However, there are barriers to overcome.’ and put ‘However’ at start of next sentence.

	- I think because you wrote ASCENDS as an acronym, all letters should be capitalized.

	- The phrases below do not seem to add to your motivation, as you have just argued that the difficulty in using big data tools requires no programming experience, and then you say that ASCENDS is a simple but power tool via CLI/GUI that requires no programming experience. You should strengthen your motivating arguments and make this more coherent.
		Most existing big data tools, systems, and methodologies have been developed for scientists (or any users) who have no…
		&
		Ascends is a toolkit that is developed to assist scientists (or any persons)…

	- Earlier versions of Ascends have been used for scientific researches —> Earlier versions of Ascends have been used for scientific research

	- For the bibliography file, there are no DOIs. Can you please include for the sources that have them. For example, your source kinney2014equitability has DOI: https://doi.org/10.1073/pnas.1309933111

	- The repository is located in the ORNLPMCP group’s account, and it is the sole contributor to the codebase. Thus, I can not confirm whether the listed authors can claim equal authorship to the submission.

**Documentation**

	- Unless I am mistaken, I see only a short markdown file that serves as an API reference. This file shows example usage, but it does not provide a proper reference to functionality per se. I recommend providing a more thorough, detailed reference for ASCEND functions.

	- I think with proper reference to ASCEND functions, it would become more clear whether a testing suite is necessary. The authors describe the codebase as a wrapper around Keras, TF, and scikit-learn, so depending on how they are ported also may influence any requisite testing functionality.

**Installation**

	- I followed the instructions using conda. Had to indeed install TF manually, as indicated. However, during `pip install ascends-toolkit`, I receive the following output, which so far hasn’t affected running of train.py.

  ERROR: Complete output from command /home/zhampel/anaconda3/envs/ascends/bin/python -u -c 'import setuptools, tokenize;__file__='"'"'/tmp/pip-install-t40_0433/minepy/setup.py'"'"';f=getattr(tokenize, '"'"'open'"'"', open)(__file__);code=f.read().replace('"'"'\r\n'"'"', '"'"'\n'"'"');f.close();exec(compile(code, __file__, '"'"'exec'"'"'))' bdist_wheel -d /tmp/pip-wheel-t0o9dv8e --python-tag cp37:
  ERROR: running bdist_wheel
  running build
  running build_py
  creating build
  creating build/lib.linux-x86_64-3.7
  creating build/lib.linux-x86_64-3.7/minepy
  copying minepy/__init__.py -> build/lib.linux-x86_64-3.7/minepy
  Fixing build/lib.linux-x86_64-3.7/minepy/__init__.py
  Skipping optional fixer: buffer
  Skipping optional fixer: idioms
  Skipping optional fixer: set_literal
  Skipping optional fixer: ws_comma
  Fixing build/lib.linux-x86_64-3.7/minepy/__init__.py
  Skipping optional fixer: buffer
  Skipping optional fixer: idioms
  Skipping optional fixer: set_literal
  Skipping optional fixer: ws_comma
  running build_ext
  Traceback (most recent call last):
    File "<string>", line 1, in <module>
    File "/tmp/pip-install-t40_0433/minepy/setup.py", line 55, in <module>
      cmdclass = {'build_ext': build_ext_custom}
    File "/home/zhampel/anaconda3/envs/ascends/lib/python3.7/site-packages/setuptools/__init__.py", line 145, in setup
      return distutils.core.setup(**attrs)
    File "/home/zhampel/anaconda3/envs/ascends/lib/python3.7/distutils/core.py", line 148, in setup
      dist.run_commands()
    File "/home/zhampel/anaconda3/envs/ascends/lib/python3.7/distutils/dist.py", line 966, in run_commands
      self.run_command(cmd)
    File "/home/zhampel/anaconda3/envs/ascends/lib/python3.7/distutils/dist.py", line 985, in run_command
      cmd_obj.run()
    File "/home/zhampel/anaconda3/envs/ascends/lib/python3.7/site-packages/wheel/bdist_wheel.py", line 192, in run
      self.run_command('build')
    File "/home/zhampel/anaconda3/envs/ascends/lib/python3.7/distutils/cmd.py", line 313, in run_command
      self.distribution.run_command(command)
    File "/home/zhampel/anaconda3/envs/ascends/lib/python3.7/distutils/dist.py", line 985, in run_command
      cmd_obj.run()
    File "/home/zhampel/anaconda3/envs/ascends/lib/python3.7/distutils/command/build.py", line 135, in run
      self.run_command(cmd_name)
    File "/home/zhampel/anaconda3/envs/ascends/lib/python3.7/distutils/cmd.py", line 313, in run_command
      self.distribution.run_command(command)
    File "/home/zhampel/anaconda3/envs/ascends/lib/python3.7/distutils/dist.py", line 985, in run_command
      cmd_obj.run()
    File "/tmp/pip-install-t40_0433/minepy/setup.py", line 9, in run
      import numpy
  ModuleNotFoundError: No module named 'numpy'
  ERROR: Failed building wheel for minepy
