### 2019-08-29 #1656 @terrytangyuan
@openjournals/joss-eics Any suggestions on this question?
> Please guide us on how we can address the following comment:
"the repository is located in the ORNLPMCP group’s account, and it is the sole contributor to the codebase. Thus, I can not confirm whether the listed authors can claim equal authorship to the submission." - We co-own the repository accounts as a team.
