### 2019-08-29 #1656 @labarba
Well, nothing can be done now to reconstruct the lost history in this project. But I want to strongly encourage the authors to adopt a traditional GitHub workflow: instead of using a `@ornlpmcp` group _account_ (you are sharing a password?? ugh!), you would create a GitHub _organization_, to which you add the various individual researchers' GitHub usernames as collaborators. It is highly unusual for more than one person to share an account like this, and it breaks many of the GitHub features: you can't `git blame` to find who in your team made a particular change, you can't assign people to issues, or even know who posted on the issue tracker.
