### 2019-09-13 #1656 @zhampel
@ornlpmcp

**Paper**
I still think there needs to be work done on the paper. I think some statements are too general

    ... so many success stories in a wide range of areas, especially in industry.

What kind of success/metric for success are you describing, and why is that relevant to your code? Are you purporting to offer the same success?

And some need better wording, for example,

    Users still need to understand what they want to do but not necessarily how it works.

I think this kind of statement should be left out, especially given that the next sentence describes what the intention of ASCENDS is.

**Usage**
I have been able to reproduce the training examples outlined in the README. However, I think that a complete classification example should be outlined in your api reference. That is, unless the calls are nearly identical to what has been shown in the regression example. Otherwise, I consider this an incomplete documentation of usage.

**GUI**
I have not been able to reproduce the GUI as shown in the README. After running the `ascends_server.py` script, I merely get the following output:
```
$ python ascends_server.py
Using TensorFlow backend.

 * ASCENDS: Advanced data SCiENce toolkit for Non-Data Scientists
 * Web Server ver 0.1

 programmed by Matt Sangkeun Lee (lees4@ornl.gov)
 please go to : http://localhost:7777/
```

Yet, when I try to go to the localhost page, I simply get a `This site can't be reached` error message (ERR_CONNECTION_REFUSED). Is there anyway to debug this to test the GUI?
