+++ paper.md
@@ -27,9 +27,9 @@ bibliography: paper.bib
 
 # Summary
 
-Advances in big data technology, artificial intelligence, and machine learning have created so many success stories in a wide range of areas, especially in industry. These success stories have been motivating scientists who study physics, chemistry, materials, medicine and many other subjects, to explore a new pathway of utilizing big data tools for their scientific activities. However, most existing big data tools, systems, and methodologies have been developed for programming experts but not for scientists (or any users) who have no or little knowledge of programming. 
+Recently, advances in machine learning and artificial intelligence have been playing more and more critical roles in a wide range of areas. For the last several years, industries have shown that how learning from data, identifying patterns and making decisions with minimal human intervention can be extremely useful to their business (e.g., image classification, recommending a product to a customer, finding friends in a social network, predicting customers actions, etc.). These success stories have been motivating scientists who study physics, chemistry, materials, medicine, and many other subjects, to explore a new pathway of utilizing machine learning techniques like regression and classification for their scientific activities. However, most existing machine learning tools, systems, and methodologies have been developed for programming experts but not for scientists (or any users) who have no or little knowledge of programming. 
 
-ASCENDS is a toolkit that is developed to assist scientists (or any persons) who want to use their data for machine learning tasks. Users still need to understand what they want to do but not necessarily how it works. ASCENDS provides a set of simple but powerful tools for non-data scientists to be able to intuitively perform various advanced data analysis and machine learning techniques with simple interfaces (a command-line interface and a web-based GUI). ASCENDS has been implemented by wrapping around open-source software including Keras [@gulli2017deep], tensorflow [@abadi2016tensorflow], scikit-learn [@pedregosa2011scikit].
+ASCENDS is a toolkit that is developed to assist scientists (or any persons) who want to use their data for machine learning tasks, more specifically, correlation analysis, regression, and classification. ASCENDS does not require programming skills.  Instead, it provides a set of simple but powerful CLI (Command Line Interface) and GUI (Graphic User Interface) tools for non-data scientists to be able to intuitively perform advanced data analysis and machine learning techniques. ASCENDS has been implemented by wrapping around open-source software including Keras [@gulli2017deep], TensorFlow [@abadi2016tensorflow], and scikit-learn [@pedregosa2011scikit].
 
 -![Using Ascends via its web-based graphic user interface](./logo/web-ui.png)
 
