+++ api_reference.md
@@ -1,19 +1,130 @@
 
 
-```python
+```
 import numpy as np
 import ascends as asc
 import keras
 import ast
+from sklearn.metrics import classification_report
+import pickle
 ```
 
 
-```python
-# Ascends-toolkit was developed to be used via command-line interface or web-based interface; however, if needed,
+```
+# 1. Regression API reference
+
+# * NOTE: Ascends-toolkit was developed to be used via command-line interface or web-based interface; however, if needed,
+# users may use ascends-toolkit's API. The following shows an example of performing a classification task using 
+# the core ascends-toolkit APIs. 
+
+# You need to download the example data file from https://raw.githubusercontent.com/pandas-dev/pandas/master/pandas/tests/data/iris.csv
+# and save it under data folder
+
+csv_file = 'data/iris.csv'
+cols_to_remove = []
+target_col = 'Name'
+input_col = None
+
+# Classifier will need a mapping between categorical values to numerical values
+mapping = {'Name': {'Iris-setosa':0, 'Iris-versicolor':1, 'Iris-virginica':2}}
+
+# Load data from csv file
+# A standard csv file can be loaded and shuffled as follows
+
+data_df, x_train, y_train, header_x, header_y = asc.data_load_shuffle(csv_file, input_col, cols_to_remove, target_col, map_all = mapping, random_state = 0)
+```
+
+
+```
+# check if data is loaded
+data_df[:10]
+```
+
+
+```
+# Generating a default model parameters
+model_parameters = asc.default_model_parameters_classifier() 
+```
+
+
+```
+model_type = 'RF'
+scaler_option = 'StandardScaler' # scaler option can be 'False','StandardScaler','Normalizer','MinMaxScaler','RobustScaler'
+num_of_folds = 5
+model = asc.define_model_classifier(model_type, model_parameters, x_header_size = x_train.shape[1])   
+```
+
+
+```
+# scikit-learn's classification report can be used to understand the accuracy of the trained model
+predictions, actual_values = asc.train_and_predict(model, x_train, y_train, scaler_option=scaler_option, num_of_folds=num_of_folds)
+accuracy = asc.evaluate_classifier(predictions, actual_values)
+print("")
+print("* Classification Report")
+print(classification_report(actual_values, predictions))
+```
+
+    
+    * Classification Report
+                  precision    recall  f1-score   support
+    
+             0.0       1.00      1.00      1.00        50
+             1.0       0.94      0.92      0.93        50
+             2.0       0.92      0.94      0.93        50
+    
+       micro avg       0.95      0.95      0.95       150
+       macro avg       0.95      0.95      0.95       150
+    weighted avg       0.95      0.95      0.95       150
+    
+
+
+
+```
+asc.train_and_save_classifier(model, "model.pkl", model_type
+                            , input_cols=header_x, target_col=header_y
+                            , x_train=x_train, y_train=y_train, scaler_option=scaler_option, path_to_save = '.', accuracy=accuracy)
+```
+
+    * Training initiated ..
+    * Training done.
+    * Trained model saved to file: model.pkl
+
+
+
+```
+# You can load the saved model by using pickle package
[…]
