### 2019-10-24 #1656 @ornlpmcp
@zhampel

Thank you very much for your reviews. They are indeed very helpful.

[Paper]
I modified the paper.md addressing your comments. I try to remove sentences that are too general and added some more examples.

[Usage]

The main purpose/contribution of this tool is to provide a user-friendly interface (CLI/GUI) so that users can do machine learning without programming efforts. So, it is not our main intention that users write codes using the APIs. In fact, it will be more efficient to use scikit-learn and keras directly. However, as you addressed, I added the API use case reference for both regression and classification so that users can use them in case they need. I checked in the api_reference.md and notebook/reference.ipynb for user reference.

[GUI]
Unfortunately, I could not replicate the error. Would you please kindly test if the port 7777 is open or not conflicted. Do you see any error messages in the console?

Please advise. I am happy to move forward.

Dear. @jrbourbeau It would be great if you could review this work. I will respond as soon as I can.
Big thanks,
Thank you for helping us too @terrytangyuan
