### 2019-11-18 #1656 @ornlpmcp
@terrytangyuan please help us to move forward. I am willing to revise the manuscript/code as advised.
