### 2019-11-20 #1656 @zhampel
@ornlpmcp

**README**

Can you please remove unnecessary dollar signs `$` for examples of single lines? They useful when displaying full outputs based on runs, however, for single lines, they are unnecessary and a bit of a nuisance for copy/paste purposes ;)

There is a erroneous slash in the following line:
`Copy the following text into a new file and save it to data\iris_test_input.csv.`
Please change to
`Copy the following text into a new file and save it to data/iris_test_input.csv.`

**Web-Based GUI**
I no longer have the port issues I mentioned before. Thanks for addressing those.

I had to do manually `pip install` various packages for running the GUI:
`minepy, matplotlib, np_utils, tornado`
Should these not be simply installed with a reqs file or similar? This would preferred over getting `No module named 'xx'` a few times when running `python ascends_server.py`, and then `pip install 'xx'`
