### 2019-11-25 #1656 @ornlpmcp
@zhampel Thanks for your review. I updated the README file accordingly.
Also, if you activated the ascends anaconda environment and if you have done installing via 'pip install ascends-toolkit' all of those packages should have been installed.

You need to install those packages separately if you run the server without activating the Anaconda environment. I added this explanation in the README file.

@jrbourbeau There is no 'testing' module; however I believe the Jupyter notebook file under /nootbook/reference.ipynb can be used for testing if the software is properly installed and run.

Thank you very much all. I would really appreciate it if you could help us move forward.
@terrytangyuan
