### 2019-11-25 #1656 @terrytangyuan
@ornlpmcp I think @jrbourbeau's point is that the software needs to be **automatically** and **continuously** tested in services like [Travis CI](https://travis-ci.com/) or [CircleCI](https://circleci.com/) whenever a new code change is introduced. A notebook would not be sufficient.
