### 2019-11-25 #1656 @kthyng
@terrytangyuan Automatic/continuous testing is not a requirement — recommended but not required. Some way to verify that the code is running correctly is required. If a notebook is provided and the way to verify that the notebook results are correct is clearly included, then that should be sufficient.
