### 2019-11-25 #1656 @terrytangyuan
@kthyng Thanks for the clarification. Sorry my point is not that this is required but to clarify the reviewer's feedback. In the meantime, @ornlpmcp it would be good to document the manual testing instructions.
