### 2019-11-25 #1656 @ornlpmcp
@terrytangyuan The readme file includes the tutorial that users can follow step by step to verify if everything works correctly, I will also mention the jupyter notebook in the readme file so that users can use it for testing
