### 2019-12-03 ornlpmcp/ASCENDS#2 @jrbourbeau
[JOSS REVIEW] Remove artifact files and add a .gitignore file
There are several files and directories in this repository that are artifacts of some other process. For example, a log file associated with a pip installation (`pip.log`), compiled Python bytecode (`__pycache__`), Jupyter notebook checkpoints (`.ipynb_checkpoints`), etc. In my experience, these types of files are not included alongside the source code for a Python project.

I recommend adding a [`.gitignore` file](https://git-scm.com/docs/gitignore) to this repository (e.g. GitHub maintains a nice [`.gitignore` for Python projects](https://github.com/github/gitignore/blob/master/Python.gitignore) that could be used here) and removing files in this repository that are already tracked by `git`, but would be ignored based on the entries in the new `.gitignore` file.

To be clear, this is not required as part of the JOSS review, but is highly recommended.

xref https://github.com/openjournals/joss-reviews/issues/1656
