### 2019-12-03 ornlpmcp/ASCENDS#3 @jrbourbeau
[JOSS REVIEW] v0.4.0 release on GitHub
The version number for ASCENDS submitted to JOSS is v0.4.0, however there is currently no release on GitHub that I can see (https://github.com/ornlpmcp/ASCENDS/releases). Could you please publish a release to GitHub? This can be done by tagging a commit and pushing the tag to GitHub, or via the GitHub web UI by following [these instructions](https://help.github.com/en/github/administering-a-repository/creating-releases).

xref https://github.com/openjournals/joss-reviews/issues/1656
