### 2019-12-03 ornlpmcp/ASCENDS#4 @jrbourbeau
[JOSS REVIEW] List package dependencies
The packages required to install and run ASCENDS should be explicitly listed in the `README.md` (e.g. in the "Installation" section).

I see there is this list of packages in `setup.py`:

https://github.com/ornlpmcp/ASCENDS/blob/4f1fd58412048446b904befd6a32c4289b1df9e9/setup.py#L21-L29

Is this list up-to-date? If so, that should be sufficient

xref https://github.com/openjournals/joss-reviews/issues/1656
