### 2019-12-03 ornlpmcp/ASCENDS#5 @jrbourbeau
[JOSS REVIEW] Add tests
Running the current [`notebook/reference.ipynb`](https://github.com/ornlpmcp/ASCENDS/blob/4f1fd58412048446b904befd6a32c4289b1df9e9/notebook/reference.ipynb) Jupyter notebook requires `ascends` to be installed, but (to my best knowledge) doesn't test any of the actual output of the `ascends` package. So running this notebook does provide a test that `ascends` is installed, but not that any of the output it is providing is correct.

Could you add some way that I can verify this software is behaving as expected? For example, in `notebook/reference.ipynb`, you're printing lots of numbers. Could you modify things such that those print statements are turned into `assert` statements? So instead of, for example, printing the model performance, you assert that the outputted model performance is equal to some known or expected value?

This is generally very useful to convince both yourself and others that things are working properly and to ensure that future modifications to the code don't unknowingly break the code's behavior.

xref https://github.com/openjournals/joss-reviews/issues/1656
