### 2019-12-03 ornlpmcp/ASCENDS#5 @jrbourbeau
I also noticed that the data files in `reference.ipynb` (`data/iris.csv` and `data/BostonHousing.csv`) are not included in the repository and running `reference.ipynb` yields `FileNotFoundError` errors unless those `.csv` files are manually downloaded. I recommend either:

- Including instructions on how to download the required data files in the "Jupyter Notebook/API Reference/Testing" section of `README.md`

- Including those data files in the repository as to avoid manual downloading altogether

- Using `scikit-learn`'s [`load_boston`](https://scikit-learn.org/stable/modules/generated/sklearn.datasets.load_boston.html#sklearn.datasets.load_boston) and [`load_iris`](https://scikit-learn.org/stable/modules/generated/sklearn.datasets.load_iris.html#sklearn.datasets.load_iris) utilities to access those two particular datasets, which are included in `scikit-learn`.
