### 2019-12-03 ornlpmcp/ASCENDS#6 @jrbourbeau
[JOSS REVIEW] Community guidelines
Could you add how others can:

1) Contribute to the software
2) Report issues or problems with the software
3) Seek support

to `README.md`? I don't currently see this information, please feel free to correct me if I've missed something
