### 2019-12-03 ornlpmcp/ASCENDS#7 @jrbourbeau
[JOSS REVIEW] Consider using conda for installing dependencies
In the installation instructions you recommend using `conda` for installing dependencies, but then proceed to install most dependencies using `pip`. You might consider creating a `conda` environment file (see https://docs.conda.io/projects/conda/en/latest/user-guide/tasks/manage-environments.html#create-env-file-manually) to instead install all dependencies using `conda` instead of with `pip`.

Further, you could even consider adding a package to [conda-forge](https://conda-forge.org/) so that ASCENDS can itself be installed using `conda`.

To be clear, this is not required as part of the JOSS review, just a recommendation to potentially avoid installation issues that have come up during the review.

xref https://github.com/openjournals/joss-reviews/issues/1656
