### 2019-12-03 #1656 @jrbourbeau
Thanks @ornlpmcp, I've opened up a few issues in the ASCENDS issue tracker with my review comments and have provided a brief outline here. Please let me know if anything in my comments is unclear or if you have any additional questions

#### Software version

- [ ] Publish a v0.4.0 release on GitHub (https://github.com/ornlpmcp/ASCENDS/issues/3)

#### Installation

- [ ] I am getting a `gcc` compilation error when trying to build an extension for `minepy` (I think similar to what @zhampel ran into in https://github.com/openjournals/joss-reviews/issues/1656#issuecomment-522658224). In the installation instructions you say "Note: you may encounter 'failed building wheel for minepy' error, but you can ignore as long as you see the 'Successfully installed minepy' message", however I was not able to successfully install `minepy` with `pip` and had to instead install it using `conda`. Given the installation errors @zhampel and I have encountered, I'd recommend providing a `conda` environment yaml file to allow for installing the necessary dependencies using `conda` instead of `pip` (https://github.com/ornlpmcp/ASCENDS/issues/7)

- [ ] Explicitly list package dependencies (https://github.com/ornlpmcp/ASCENDS/issues/4)

#### Documentation

- [ ] Could you add to `README.md` that the available command line options are available via `python train.py --help`?

- [ ] Add `assert` statements to notebook to ensure output is as expected (https://github.com/ornlpmcp/ASCENDS/issues/5)

- [ ] Related to the paper: Is there a link to the rendered paper PDF? I see the paper markdown file over in the [GitHub repo](https://github.com/ornlpmcp/ASCENDS/blob/master/paper.md), but was the PDF compilation error here https://github.com/openjournals/joss-reviews/issues/1656#issuecomment-522196215 ever resolved?

- [ ] Add community guidelines (https://github.com/ornlpmcp/ASCENDS/issues/6)
