### 2019-12-03 aa44825
removed artifact files
