+++ ascends_toolkit.egg-info/requires.txt
@@ -1,9 +1,0 @@
-matplotlib
-numpy
-pandas
-tensorflow
-keras
-tornado
-scikit-learn
-minepy
-np_utils
