+++ ascends_toolkit.egg-info/SOURCES.txt
@@ -1,35 +1,0 @@
-ascends.py
-ascends_server.py
-classify_with_model.py
-predict_with_model.py
-setup.py
-train_classifier.py
-train_regression.py
-ascends_toolkit.egg-info/PKG-INFO
-ascends_toolkit.egg-info/SOURCES.txt
-ascends_toolkit.egg-info/dependency_links.txt
-ascends_toolkit.egg-info/requires.txt
-ascends_toolkit.egg-info/top_level.txt
-output/.DS_Store
-output/.gitignore
-static/.DS_Store
-static/common.js
-static/custom.css
-static/config/.gitignore
-static/config/creep.csv,Model=BR,Scaler=Normalizer.tuned.prop
-static/config/creep.csv,Model=KR,Scaler=Normalizer.tuned.prop
-static/config/creep.csv,Model=NET,Scaler=StandardScaler,MAE=273.66888,R2=0.8629746873805393.tuned.checkpoint.prop
-static/config/creep.csv,Model=RF,Scaler=StandardScaler.tuned.prop
-static/config/creep.csv,Model=SVM,Scaler=Normalizer.tuned.prop
-static/config/creep.csv,Model=SVM,Scaler=StandardScaler.tuned.prop
-static/img/error.png
-static/learned_models/.gitignore
-static/output/.DS_Store
-static/output/.gitignore
-static/output/ml/.gitignore
-static/output/ml/ml_result.png
-static/uploads/.gitignore
-templates/index.html
-templates/ml.html
-templates/predict.html
-templates/table.html
