### 2019-12-03 ornlpmcp/ASCENDS#2 @ornlpmcp
Thanks for your review! I removed the artifact files you mentioned and added a .gitignore file in the repository.
