### 2019-12-03 ornlpmcp/ASCENDS#2 @jrbourbeau
Great, thank you! `build/` and `dist/` are both mentioned in the `.gitignore` file and can also be removed too (those are Python package distribution artifacts). Is the `output/` directory being used anywhere? If not, I suspect it may be okay to remove that too
