### 2019-12-03 9ffd0ed
redesigned the installation process
