+++ build/scripts-3.7/train_classifier.py
@@ -1,366 +1,0 @@
-#!python -W ignore::RuntimeWarning
-# coding: utf-8
-import warnings
-import argparse
-import statistics
-import sys
-import pprint
-import keras
-import ascends as asc
-import ast
-from sklearn.metrics import classification_report
-from sklearn.metrics import confusion_matrix
-
-# This is a tool to training machine learning models for a regression (value prediction) task
-
-def clean_up_net_params(net_neuron_max, net_structure, net_l_2, net_learning_rate, net_epochs, net_dropout, net_layer_n, net_batch_size):
-    if net_neuron_max=='-1':
-        net_neuron_max = []
-    else:
-        try:
-            net_neuron_max = [int(x) for x in net_neuron_max]
-        except:
-            net_neuron_max = []
-
-    if net_structure=='Tune':
-        net_structure = None
-    else:
-        try:
-            net_structure = [int(x) for x in net_structure]
-        except:
-            net_structure = []
-
-    if net_l_2=='Tune':
-        net_l_2 = None
-    else:
-        try:
-            net_l_2 = float(net_l_2)
-        except:
-            net_l_2 = None
-    
-    if net_learning_rate=='Tune':
-        net_learning_rate = None
-    else:
-        try:
-            net_learning_rate = float(net_learning_rate)
-        except:
-            net_learning_rate = None
-    
-    if net_epochs=='Tune':
-        net_epochs = None
-    else:
-        try:
-            net_epochs = int(net_epochs)
-        except:
-            net_epochs = None
-
-    if net_dropout=='Tune':
-        net_dropout = True
-    else:
-        try:
-            net_dropout = float(net_dropout)
-        except:
-            net_dropout = True
-
-    if net_layer_n=='Tune':
-        net_layer_n = None
-    else:
-        try:
-            net_layer_n = int(net_layer_n)
-        except:
-            net_layer_n = None
-    
-    if net_batch_size=='Tune':
-        net_batch_size = None
-    else:
-        try:
-            net_batch_size = int(net_batch_size)
-        except:
-            net_batch_size = None
-
-    return net_neuron_max, net_structure, net_l_2, net_learning_rate, net_epochs, net_dropout, net_layer_n, net_batch_size
-
-def main(args):
-
-    # loading data
-    try:
-        
-        print("\n [ Data Loading ]")
-
-        csv_file = args.input_file
-        cols_to_remove = args.ignore_col
-        target_col = args.target_col
-        input_col = args.input_col
-        model_type = args.model_type
-        hyperparameter_file = asc.fix_value(args.hyperparameter_file,'str')
-        num_of_features = int(args.num_of_features)
-        num_of_folds = int(args.num_of_folds)
-        test = asc.str2bool(args.test)
-        output_file = args.output_file
-        save_auto_tune = asc.str2bool(args.save_auto_tune)
-        auto_tune = asc.str2bool(args.auto_tune)
-        auto_tune_iter = int(args.auto_tune_iter)
-        random_state = asc.fix_value(args.random_state,'int')
-        feature_selection = args.feature_selection
-        scaler_option = args.scaler
-        save_corr_chart = args.save_corr_chart
-        save_corr_report = args.save_corr_report
-        mapping = args.mapping
-        net_structure = args.net_structure
-        net_layer_n = args.net_layer_n
-        net_dropout = args.net_dropout
-        net_l_2 = args.net_l_2
-        net_learning_rate = args.net_learning_rate
-        net_epochs = args.net_epochs
-        net_batch_size = args.net_batch_size
-        net_neuron_max = args.net_neuron_max
-        net_batch_size_max = int(args.net_batch_size_max)
[…]
