+++ conda_requirements.txt
@@ -1,6 +1,0 @@
-matplotlib
-numpy
-pandas
-tensorflow
-keras
-scikit-learn
