+++ environment.yml
@@ -1,0 +1,13 @@
+name: ascends
+channels:
+  - bioconda
+dependencies:
+ - python=3.6
+ - matplotlib=3.0.3
+ - numpy=1.16.4
+ - pandas=0.24.2
+ - tensorflow=1.13.1  
+ - keras=2.2.4 
+ - scikit-learn=0.20.3
+ - tornado=6.0.2
+ - minepy=1.2.3

[…]
