### 2019-12-03 ornlpmcp/ASCENDS#2 @ornlpmcp
They are removed too. Thanks - I am working on other review items. Having an 'output' directory makes it easier for users to test code following the example. Stay tuned thanks a lot.
