### 2019-12-03 ornlpmcp/ASCENDS#4 @ornlpmcp
Now all dependencies with proper versions are in 'environment.yml'
I updated the installation process to use conda as you guided.
no pip is needed.
Thanks a lot.
