### 2019-12-03 ornlpmcp/ASCENDS#4 @ornlpmcp
I also updated README.md to mention the environment.yml and remove description on pip and pip_requirement.txt and conda_requirement.txt
