### 2019-12-03 ornlpmcp/ASCENDS#7 @ornlpmcp
I changed the installation process so now pip does not need to be installed. You can simply install all dependencies by creating a dedicated environment using environment.yml in the repository. I will look into conda-forge, but for now I want to keep it simple. Thanks again for your review.
