### 2019-12-03 ornlpmcp/ASCENDS#7 @jrbourbeau
Thanks @ornlpmcp!

> I will look into conda-forge, but for now I want to keep it simple

Sounds good. That's certainly not required, just thought it was worth mentioning as an option
