### 2019-12-04 e64740e
added test asserts in notebook
