+++ environment.yml
@@ -11,3 +11,4 @@ dependencies:
  - scikit-learn=0.20.3
  - tornado=6.0.2
  - minepy=1.2.3
+ - jupyter=1.0.0
