### 2019-12-04 ornlpmcp/ASCENDS#5 @ornlpmcp
@jrbourbeau I modified the notebook/reference.ipynb according to your suggestion
- I included all the dataset in the repository
- now dependencies include jupyter notebook for users convenience of testing
- reference.ipynb now includes asserts so the users can test if everything works correctly.
Thanks for your suggenstions.
