+++ paper.md
@@ -35,14 +35,11 @@ Recently, advances in machine learning and artificial intelligence have been pla
 
 ASCENDS is a toolkit that is developed to assist scientists (or any persons) who want to use their data for machine learning tasks, more specifically, correlation analysis, regression, and classification. ASCENDS does not require programming skills.  Instead, it provides a set of simple but powerful CLI (Command Line Interface) and GUI (Graphic User Interface) tools for non-data scientists to be able to intuitively perform advanced data analysis and machine learning techniques. ASCENDS has been implemented by wrapping around open-source software including Keras [@gulli2017deep], TensorFlow [@abadi2016tensorflow], and scikit-learn [@pedregosa2011scikit].
 
--![Using Ascends via its web-based graphic user interface](./logo/web-ui.png)
+![Using Ascends via its web-based graphic user interface](./logo/web-ui.png)
 
--![Using Ascends via its command-line interface](./logo/command-line-ui.png)
+![Using Ascends via its command-line interface](./logo/command-line-ui.png)
 
-ASCENDS provides three capabilities as follows.
-- Correlation analysis [@ezekiel1930methods]: Users can measure the correlation between input variables (X) and an output variable (Y) using various correlation coefficients including Pearson's correlation coefficient [@sedgwick2012pearson] and maximal information coefficient [@kinney2014equitability]. 
-- Training, testing, saving and using classification models [@ren2003learning]: Users can train a predictive model (mapping function) that predicts a category (Y) from input variables (X) using ASCENDS. For instance, ASCENDS can train a model for predicting whether an email is a spam or not-spam.
-- Training, testing, saving and using regression models [@darlington1990regression]: Users can train a predictive model (mapping function) that approximates a continuous output variable (y) from input variables (X) using ASCENDS. For instance, ASCENDS can train a model for predicting the value of a house based on square footage, number of bedrooms, number of cars that can be parked in its garages, number of storages.
+Users can perform three major tasks using ASCENDS as follows. First of all, users can easily perform correlation analysis [@ezekiel1930methods] using ASCENDS. ASCENDS can quantify the correlation between input variables ($X$) and an output variable ($y$) using various correlation coefficients including Pearson's correlation coefficient [@sedgwick2012pearson] and maximal information coefficient [@kinney2014equitability]. Second, users can train, test, save and utilize classification models [@ren2003learning] without any programming efforts. For instance, with ASCENDS, by executing a single command in a terminal, user can train a model for predicting whether an email is a spam or not-spam. Last, similarly, users can train, test, save and use regression models [@darlington1990regression]. For instance, ASCENDS can be used to train a model for predicting the value of a house based on square footage, number t bedrooms, number of cars that can be parked in its garages, number of storages.
 
 Earlier versions of Ascends have been used for scientific research such as [@shin2019modern] [@shin2017petascale] [@wang2019machine]. 
 
