### 2019-12-04 ornlpmcp/ASCENDS#6 @ornlpmcp
added community guidelines in the README.md thanks!
