+++ paper.md
@@ -39,7 +39,7 @@ ASCENDS is a toolkit that is developed to assist scientists (or any persons) who
 
 ![Using Ascends via its command-line interface](./logo/command-line-ui.png)
 
-Users can perform three major tasks using ASCENDS as follows. First of all, users can easily perform correlation analysis [@ezekiel1930methods] using ASCENDS. ASCENDS can quantify the correlation between input variables ($X$) and an output variable ($y$) using various correlation coefficients including Pearson's correlation coefficient [@sedgwick2012pearson] and maximal information coefficient [@kinney2014equitability]. Second, users can train, test, save and utilize classification models [@ren2003learning] without any programming efforts. For instance, with ASCENDS, by executing a single command in a terminal, user can train a model for predicting whether an email is a spam or not-spam. Last, similarly, users can train, test, save and use regression models [@darlington1990regression]. For instance, ASCENDS can be used to train a model for predicting the value of a house based on square footage, number t bedrooms, number of cars that can be parked in its garages, number of storages.
+Users can perform three major tasks using ASCENDS as follows. First of all, users can easily perform correlation analysis [@ezekiel1930methods] using ASCENDS. ASCENDS can quantify the correlation between input variables ($X$) and an output variable ($y$) using various correlation coefficients including Pearson's correlation coefficient [@sedgwick2012pearson] and maximal information coefficient [@kinney2014equitability]. Second, users can train, test, save and utilize classification models [@ren2003learning] without any programming efforts. For instance, with ASCENDS, by executing a single command in a terminal, a user can train a model for predicting whether an email is a spam or not-spam. Last, similarly, users can train, test, save and use regression models [@darlington1990regression]. For instance, ASCENDS can be used to train a model for predicting the value of a house based on square footage, number t bedrooms, number of cars that can be parked in its garages, number of storages using the provided graphic user interface in a web browser.
 
 Earlier versions of Ascends have been used for scientific research such as [@shin2019modern] [@shin2017petascale] [@wang2019machine]. 
 
