### 2019-12-04 #1656 @ornlpmcp
@terrytangyuan @zhampel @jrbourbeau

Dear reviewers, I've addressed all of your review items as far as I understand.
Please advise, if there's anything else I need to work on. Here are the major updates.

- I fixed paper.md to deal with the compile error.
- I addressed all issues mentioned by @jrbourbeau
- I added testing in the jupyter notebook
- I simplified the process of installation
- I added community guidelines and additional explanations on several things
- I added gitignore and removed artifact files
- v0.4.0 has been released in github
- I modified the paper.md to clarify explanations

Thank you very much and help us to move forward.

Sangkeun (Matt) Lee
