+++ ascends.py
@@ -1,10 +1,11 @@
 # coding: utf-8
-
 from __future__ import print_function
 # In[1]:
 import warnings
 warnings.filterwarnings('ignore')
 warnings.simplefilter("ignore")
+import os
+os.environ['KMP_DUPLICATE_LIB_OK']='True'
 
 from keras import regularizers
 from keras.datasets import mnist
@@ -65,6 +66,8 @@ from minepy import MINE
 from sklearn.linear_model import LogisticRegression
 from os import path
 from pathlib import PurePath
+from tensorflow.python.util import deprecation
+deprecation._PRINT_DEPRECATION_WARNINGS = False
 
 def save_test_data(predictions, actual_values, filename):
     df_predictions = pd.DataFrame(predictions, columns=['predictions'])
@@ -188,75 +191,111 @@ def data_load_shuffle(csv_file, input_col, cols_to_remove, target_col, random_st
     
     return data_df, x_train, y_train, header_x, header_y
 
-def correlation_analysis_all(data_df, target_col, top_k=10, file_to_save = None, save_chart = None):
-    print("* correlation_analysis_all")
-    pcc = data_df.corr()[target_col]
+def correlation_analysis_all(data_df, target_col, top_k=10, file_to_save = None, save_chart = None, only_pcc= False, feature_selection_file = None):
+    
+    if feature_selection_file == None:
+        print("* correlation_analysis_all")
+        pcc = data_df.corr()[target_col]
 
-    if(len(pcc)<top_k):
-        top_k=len(pcc)
+        if(len(pcc)<top_k):
+            top_k=len(pcc)
+        print("Computing PCC, PCC_SQRT ..")
+        pcc = pcc.sort_values(ascending = False).dropna()
+        pcc = pcc.rename("PCC")
+        try:
+            del pcc[target_col]
+        except:
+            pass
+        pcc_sqrt = pcc.apply(lambda x: np.sqrt(x* x))
+        pcc_sqrt = pcc_sqrt.sort_values(ascending = False).dropna()
+        pcc_sqrt = pcc_sqrt.rename("PCC_SQRT")
+        MICs = []
+        MASs = []
+        MEVs = []
+        MCNs = []
+        MCN_generals = []
+        GMICs = []
+        TICs = []
+        print("Computing all other metrics ..")
+        if only_pcc==False or only_pcc=='False':
+            for col in data_df.columns:
+                print(" - computing for ", col, "...")
+                if col!=target_col:
+                    x = data_df[col].values
+                    y = data_df[target_col].values
+                    mine = MINE()
+                    mine.compute_score(x,y)
+                    MICs.append((col,mine.mic()))
+                    MASs.append((col,mine.mas()))
+                    MEVs.append((col,mine.mev()))
+                    MCNs.append((col,mine.mcn(0)))
+                    MCN_generals.append((col,mine.mcn_general()))
+                    GMICs.append((col,mine.gmic()))
+                    TICs.append((col,mine.tic())) 
+                
+        top_k_pcc = list(pcc.keys())[:top_k]
+        top_k_pcc_sqrt = list(pcc_sqrt.keys())[:top_k]
+        top_k_mic = [tup[0] for tup in sorted(MICs, key=lambda tup: tup[1], reverse = True)[:top_k]]
[…]
