+++ ascends.py
@@ -546,7 +546,7 @@ def str2bool(v):
   return v.lower() in ("yes", "true", "t", "1")
 
 
-def cross_val_predict_net_classifier(model, x_train, y_train, epochs=1000, batch_size=8, verbose = 0, scaler_option='StandardScaler', num_of_folds = 5, num_of_class = 2, force_to_proceed = False, accuracy_threshold = 0.5):
+def cross_val_predict_net_classifier(model, x_train, y_train, epochs=1000, batch_size=8, verbose = 0, scaler_option='StandardScaler', num_of_folds = 5, num_of_class = 2, force_to_proceed = False, accuracy_threshold = 0.5, fast_tune = True):
     
     x_trains, y_trains, x_tests, y_tests = split_data(x_train, y_train, num_of_folds=num_of_folds)
     
@@ -586,11 +586,14 @@ def cross_val_predict_net_classifier(model, x_train, y_train, epochs=1000, batch
 
         predictions_total+=list(predictions)
         actual_values_total+=list(actual_values)
+        if fast_tune==True:
+            print("* Fast tuning enabled. so we only test 1 fold, and move on ..")
+            break
 
     return np.array(predictions_total), np.array(actual_values_total)
 
 
-def cross_val_predict_net(model, x_train, y_train, epochs=1000, batch_size=8, verbose = 0, scaler_option='StandardScaler', num_of_folds = 5, force_to_proceed= False):
+def cross_val_predict_net(model, x_train, y_train, epochs=1000, batch_size=8, verbose = 0, scaler_option='StandardScaler', num_of_folds = 5, force_to_proceed= False, fast_tune=True):
     
     x_trains, y_trains, x_tests, y_tests = split_data(x_train, y_train, num_of_folds=num_of_folds)
     
@@ -626,6 +629,9 @@ def cross_val_predict_net(model, x_train, y_train, epochs=1000, batch_size=8, ve
 
         predictions_total+=list(predictions)
         actual_values_total+=list(actual_values)
+        if fast_tune==True:
+            print("* Fast tuning enabled. so we only test 1 fold, and move on ..")
+            break
 
     return np.array(predictions_total), np.array(actual_values_total)
 
@@ -678,7 +684,7 @@ def train_and_save_net_classifier(model, tag, input_cols, target_col, x_train, y
         
         print('* Model has not been evaluated. Evaluation initiated via %d-fold cross validation'%(num_of_folds))
         
-        predictions, actual_values = cross_val_predict_net_classifier(model, epochs=epochs, batch_size=batch_size, x_train = x_train, y_train = y_train, verbose = 0, scaler_option = scaler_option, num_of_folds = num_of_folds, num_of_class = num_of_class)
+        predictions, actual_values = cross_val_predict_net_classifier(model, epochs=epochs, batch_size=batch_size, x_train = x_train, y_train = y_train, verbose = 0, scaler_option = scaler_option, num_of_folds = num_of_folds, num_of_class = num_of_class, fast_tune = False)
         accuracy = evaluate_classifier(predictions, actual_values)
 
     x_train_, scale = rescale_x(scaler_option, x_train)
@@ -712,7 +718,7 @@ def train_and_save_net(model, tag, input_cols, target_col, x_train, y_train, sca
         
         print('* Model has not been evaluated. Evaluation initiated via %d-fold cross validation'%(num_of_folds))
         
[…]
