### 2019-12-10 #1656 @zhampel
@terrytangyuan Yes, I'll chip away at this during this week, as NeurIPS is taking most of my time ;) I'll get back by the end of the week.
