### 2020-01-02 b195cf9
fix a bug (manual layer_n set)
