### 2020-01-03 634d3f8
fixed a bug regarding learning rate
