+++ ascends.py
@@ -5,6 +5,9 @@ import warnings
 warnings.filterwarnings('ignore')
 warnings.simplefilter("ignore")
 import os
+import numpy as np
+import tensorflow as tf
+import random as rn
 os.environ['KMP_DUPLICATE_LIB_OK']='True'
 
 from keras import regularizers
@@ -54,7 +57,6 @@ import math
 import matplotlib
 matplotlib.use("TkAgg")
 from matplotlib import pyplot as plt
-import numpy as np
 import os
 import pandas as pd
 import pickle
@@ -67,6 +69,7 @@ from sklearn.linear_model import LogisticRegression
 from os import path
 from pathlib import PurePath
 from tensorflow.python.util import deprecation
+
 deprecation._PRINT_DEPRECATION_WARNINGS = False
 
 def save_test_data(predictions, actual_values, filename):
@@ -586,6 +589,7 @@ def cross_val_predict_net_classifier(model, x_train, y_train, epochs=1000, batch
 
         predictions_total+=list(predictions)
         actual_values_total+=list(actual_values)
+        
         if fast_tune==True:
             print("* Fast tuning enabled. so we only test 1 fold, and move on ..")
             break
@@ -1148,8 +1152,20 @@ def split_data(x_train, y_train, num_of_folds=5):
         
     return x_trains, y_trains, x_tests, y_tests
 
-def net_define(params = [8, 8, 8], layer_n = 3, input_size = 16, dropout=0, l_2=0.01, optimizer='adam'):
-        
+def net_define(params = [8, 8, 8], layer_n = 3, input_size = 16, dropout=0, l_2=0.01, optimizer='adam', random_state = 0):
+    
+    os.environ['PYTHONHASHSEED'] = str(random_state)
+
+    # Setting the seed for numpy-generated random numbers
+    np.random.seed(random_state)
+
+    # Setting the seed for python random numbers
+    rn.seed(random_state)
+
+    # Setting the graph-level random seed.
+    tf.set_random_seed(random_state)
+
+    
     if len(params)!=layer_n or layer_n<1:
         return None
     
@@ -1168,7 +1184,18 @@ def net_define(params = [8, 8, 8], layer_n = 3, input_size = 16, dropout=0, l_2=
     
     return model
 
-def net_define_classifier(params = [8, 8, 8], layer_n = 3, num_of_class = 2, input_size = 16, dropout=0, l_2=0.01, optimizer='adam'):
+def net_define_classifier(params = [8, 8, 8], layer_n = 3, num_of_class = 2, input_size = 16, dropout=0, l_2=0.01, optimizer='adam', random_state = 0):
+    
+    os.environ['PYTHONHASHSEED'] = str(random_state)
+
+    # Setting the seed for numpy-generated random numbers
+    np.random.seed(random_state)
+
+    # Setting the seed for python random numbers
+    rn.seed(random_state)
+
+    # Setting the graph-level random seed.
+    tf.set_random_seed(random_state)
     
     if len(params)!=layer_n or layer_n<1:
         return None
@@ -1202,7 +1229,7 @@ def evaluate_net(model, x_train, y_train, x_test, y_test, epochs=1000, batch_siz
     
     return score, history
 
-def net_tuning_classifier(x_train, y_train, num_of_class = 2, tries = 10, lr = None, layer = None, params=None, epochs=None, batch_size=None, dropout=None, l_2 = None, neuron_max=[64, 64, 64], batch_size_max=32, layer_min=1, layer_max=3, dropout_max=0.2, scaler_option='StandardScaler', default_neuron_max=32, checkpoint = None, num_of_folds=5, fast_tune = True):
[…]
