### 2020-01-09 461f4a9
minor bug fixes
