+++ train.py
@@ -16,6 +16,20 @@ from pathlib import PurePath
 import numpy as np
 import tensorflow as tf
 import random as rn
+from tensorflow.python.client import device_lib
+from keras import backend as K
+import ascends as asc
+
+os.environ['TF_CPP_MIN_LOG_LEVEL'] = '3'
+
+#sess = tf.Session(config=tf.ConfigProto(log_device_placement=True))
+
+#config = tf.ConfigProto( device_count = {'GPU': 1 , 'CPU':1 } )
+#sess = tf.Session(config=config)
+#keras.backend.set_session(sess)
+
+print(device_lib.list_local_devices())
+print("GPUs for Keras:",K.tensorflow_backend._get_available_gpus())
 
 # This is a tool for training machine learning models for a regression (value prediction) task
 
