### 2020-01-17 dc704c1
fixed a major bug - neural network for classification
