+++ paper.bib
@@ -6,7 +6,7 @@
   pages={321--330},
   year={2019},
   publisher={Elsevier},
-  doi={https://doi.org/10.1016/j.actamat.2019.02.017}
+  doi={10.1016/j.actamat.2019.02.017}
 }
 
 @article{shin2017petascale,
@@ -18,7 +18,7 @@
   pages={828--838},
   year={2017},
   publisher={Taylor \& Francis},
-  doi={https://doi.org/10.1080/14686996.2017.1371559}
+  doi={10.1080/14686996.2017.1371559}
 }
 
 @article{wang2019machine,
@@ -27,7 +27,7 @@
   journal={Advanced Theory and Simulations},
   year={2019},
   publisher={Wiley Online Library},
-  doi={https://doi.org/10.1002/adts.201800196}
+  doi={10.1002/adts.201800196}
 }
 
 @book{gulli2017deep,
@@ -64,7 +64,7 @@
   pages={e4483},
   year={2012},
   publisher={British Medical Journal Publishing Group},
-  doi={https://doi.org/10.1136/bmj.e4483}
+  doi={10.1136/bmj.e4483}
 }
 
 @article{kinney2014equitability,
@@ -76,7 +76,7 @@
   pages={3354--3359},
   year={2014},
   publisher={National Acad Sciences},
-  doi={https://doi.org/10.1073/pnas.1309933111}
+  doi={10.1073/pnas.1309933111}
 }
 
 @article{ezekiel1930methods,
@@ -93,7 +93,7 @@
   pages={10},
   year={2003},
   organization={IEEE},
-  doi={https://doi.org/10.1109/ICCV.2003.1238308}
+  doi={10.1109/ICCV.2003.1238308}
 }
 
 @book{darlington1990regression,
