### 2020-02-04 #1656 @terrytangyuan
@openjournals/joss-eics The paper looks good now. Could you take it from here?
