### 2020-02-04 #1656 @kthyng
I am taking over the submission from here. I've been looking through your paper and have found errors. Please consider the pull requests: [#10](https://github.com/ornlpmcp/ASCENDS/pull/10) and [#9](https://github.com/ornlpmcp/ASCENDS/pull/9).

I have some remaining concerns about the references:
* MaTerialS: is this properly written with strange capitalization?
* two of the references have "et al" — is there a reason you are not including the full author list?
