+++ paper.bib
@@ -12,7 +12,7 @@
 @article{shin2017petascale,
   title={Petascale supercomputing to accelerate the design of high-temperature alloys},
   author={Shin, Dongwon and Lee, Sangkeun and Shyam, Amit and Haynes, J Allen},
-  journal={Science and Technology of advanced MaTerialS},
+  journal={Science and Technology of advanced Materials},
   volume={18},
   number={1},
   pages={828--838},
