### 2020-02-04 #1656 @ornlpmcp
@kthyng I fixed capitalization, merged pull requests, and added full list of authors.
Thanks a lot.
