+++ paper.bib
@@ -9,6 +9,17 @@
   doi={10.1016/j.actamat.2019.02.017}
 }
 
+@article{reshef2011detecting,
+  title={Detecting novel associations in large data sets},
+  author={Reshef, David N and Reshef, Yakir A and Finucane, Hilary K and Grossman, Sharon R and McVean, Gilean and Turnbaugh, Peter J and Lander, Eric S and Mitzenmacher, Michael and Sabeti, Pardis C},
+  journal={science},
+  volume={334},
+  number={6062},
+  pages={1518--1524},
+  year={2011},
+  publisher={American Association for the Advancement of Science}
+}
+
 @article{shin2017petascale,
   title={Petascale supercomputing to accelerate the design of high-temperature alloys},
   author={Shin, Dongwon and Lee, Sangkeun and Shyam, Amit and Haynes, J Allen},
