+++ paper.bib
@@ -12,7 +12,7 @@
 @article{reshef2011detecting,
   title={Detecting novel associations in large data sets},
   author={Reshef, David N and Reshef, Yakir A and Finucane, Hilary K and Grossman, Sharon R and McVean, Gilean and Turnbaugh, Peter J and Lander, Eric S and Mitzenmacher, Michael and Sabeti, Pardis C},
-  journal={science},
+  journal={Science},
   volume={334},
   number={6062},
   pages={1518--1524},
@@ -60,7 +60,7 @@
 @article{pedregosa2011scikit,
   title={Scikit-learn: Machine learning in {Python}},
   author={Pedregosa, Fabian and Varoquaux, Ga{\"e}l and Gramfort, Alexandre and Michel, Vincent and Thirion, Bertrand and Grisel, Olivier and Blondel, Mathieu and Prettenhofer, Peter and Weiss, Ron and Dubourg, Vincent and Vanderplas, Jake and Passos, Alexandre and Cournapeau, David and Brucher, Matthieu and Perrot, and Matthieu and Duchesnay, {\'{E}}douard},
-  journal={Journal of machine learning research},
+  journal={Journal of Machine Learning Research},
   volume={12},
   number={Oct},
   pages={2825--2830},
