### 2020-02-05 #1656 @kthyng
@ornlpmcp Sorry about the delay but I can't seem to get your changes to the bib file to come through to the generated file. I tried a different browser and had the same problem, though when I went to edit the bib files I then was able to see that you had updated them. I'm still seeing "et al"s in three entries, for example.

@openjournals/dev do you know of a way to force the new code to be used?
