### 2020-02-05 #1656 @kthyng
Ok ok, I see — there was some delay at first, but I think what's happening now is our bib renderer converts the list of authors to "et al" after some number of authors.
